requirements.txt
discord.py
pymongo
emoji
pytz